Reorganization, screenshots, cleanup, notes to self
