+++ config/api.config.ts
@@ -1,3 +1,3 @@
 const isDevEnvironment = process.env.NODE_ENV !== 'production';
 
-export const API_URL = isDevEnvironment ? 'http://localhost:3000' : 'http://my.productionenvironment.com/';
+export const API_URL = isDevEnvironment ? 'http://localhost:3000' : 'http://localhost:3000';
