+++ package.json
@@ -14,8 +14,6 @@
   "dependencies": {
     "@deck.gl/core": "^8.5.3",
     "@deck.gl/extensions": "^8.5.3",
-    "@material-ui/core": "^4.12.3",
-    "@material-ui/icons": "^4.11.2",
     "@reduxjs/toolkit": "^1.6.1",
     "deck.gl": "^8.5.3",
     "luxon": "^2.0.2",
@@ -31,7 +29,7 @@
     "xml2js": "^0.4.23"
   },
   "devDependencies": {
-    "@danmarshall/deckgl-typings": "^4.9.3",
+    "@danmarshall/deckgl-typings": "^4.9.7",
     "@types/enzyme": "^3.10.9",
     "@types/jest": "^26.0.24",
     "@types/luxon": "^1.27.1",
