+++ src/styles/components/map/Map.module.scss
@@ -1,3 +1,4 @@
 .map {
-
+  margin: 0;
+  height: 100vh;
 }
