+++ src/styles/components/panel/Details.module.scss
@@ -2,7 +2,7 @@
   background-color: #111;
   color: #eee;
   width: 300px;
-  height: 725px;
+  height: 100;
   opacity: 0.8;
   border-radius: 10px;
   padding: 10px;
