+++ src/styles/components/panel/RouteDetail.module.scss
