+++ src/styles/components/panel/Routes.module.scss
