+++ src/styles/components/panel/StationDetails.module.scss
@@ -27,8 +27,11 @@
 }
 
 .trainsContainer {
-  height: 455px;
-  overflow-y: auto;
+  display: flex;
+  flex-direction: column;
+  overflow-y: scroll;
+  height: 100%;
+  max-height: 500px;
 }
 
 .trains {
